requirements.txt
cachelib==0.10.2
click==8.1.3
Flask==2.2.3
Flask-Login==0.6.2
Flask-Session==0.4.0
Flask-WTF==1.1.1
gunicorn==20.0.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.33
protobuf==3.20.3
psycopg2-binary==2.9.6
Werkzeug==2.2.3
WTForms==3.0.1
